kernel: debug: move UartDebugWriter to capsules/system
