+++ capsules/system/src/debug_writer/mod.rs
@@ -1,0 +1,7 @@
+// Licensed under the Apache License, Version 2.0 or the MIT License.
+// SPDX-License-Identifier: Apache-2.0 OR MIT
+// Copyright Tock Contributors 2025.
+
+//! Implementations of `DebugWriter` for kernel-wide debug printing.
+
+pub mod uart_debug_writer;
