+++ capsules/system/src/lib.rs
@@ -5,6 +5,7 @@
 #![forbid(unsafe_code)]
 #![no_std]
 
+pub mod debug_writer;
 pub mod process_checker;
 pub mod process_policies;
 pub mod process_printer;
